[refact] move ColorsType to types.ts
