+++ app/types.ts
@@ -2,3 +2,9 @@ export interface MsgType {
     is_receive: boolean
     message: string
 }
+
+export interface ColorsType {
+    background: string
+    is_receive: string
+    is_not_receive: string
+}
